src/requirements.txt
tweepy
pandas
configparser
pymongo
pip
time
datetime
os
numpy
spacy.lang.en
nltk
nltk.stem.snowball
en_core_web_sm
emoji
nltk.sentiment.vader
poetry
dvc
dvc-gdrive
csv
dns
ssl